requirements.txt
Django==3.2.5
python-dotenv==0.18.0
Pillow==8.3.1
django-admin-sortable2==1.0